// src/app/models/bodega.model.ts
export class BodegaModel {
    bodegaId?:            number;
    bodegaNombre?:        string;
    bodegaDireccion?:     string;
    bodegaTelefono?:      string;
    sucursalId?:          number;
    sucursalDescripcion?: string;
    estadoId?:            number;
    estadoDescripcion?:   string;
    fechaHoraReg?:        Date;
    usuIdReg?:            number;
    usuRegName?:          string;
}