// src/app/models/empresa.model.ts
export class EmpresaModel {
    empresaId?:              number;
    empresaRuc?:             string;
    empresaNombre?:          string;
    empresaRazon?:           string;
    empresaDireccionMatriz?: string;
    empresaTelefonoMatriz?:  string;
    ciudadId?:               number;
    ciudadDescripcion?:      string;
    estadoId?:               number;
    estadoDescripcion?:      string;
    fechaHoraReg?:           Date;
    usuIdReg?:               number;
    usuRegName?:             string;
}