// src/app/models/producto.model.ts
export class ProductoModel {
    prodId?:               number;
    prodDescripcion?:      string;
    prodUltPrecio?:        number;
    categoriaId?:          number;
    categoriaDesripcion?:  string;
    empresaId?:            number;
    empresaDescripcion?:   string;
    proveedorId?:          number;
    proveedorDescripcion?: string;
    estadoId?:             number;
    estadoDescripcion?:    string;
    fechaHoraReg?:         Date;
    usuIdReg?:             number;
    usuRegName?:           string;
}