// src/app/models/sucursal.model.ts
export class SucursalModel {
    sucursalId?:            number;
    sucursalRuc?:           string;
    sucursalNombre?:        string;
    sucursalDireccion?:     string;
    sucursalTelefono?:      string;
    codEstablecimientoSri?: string;
    empresaId?:             number;
    empresaDescripcion?:    string;
    estadoId?:              number;
    estadoDescripcion?:     string;
    fechaHoraReg?:          Date;
    usuIdReg?:              number;
    usuRegName?:            string;
}